TASK: Rename Neos.PropertyAccess and add TypeHandling Utility
The ``Neos.PropertyAccess`` package has become ``Neos.ObjectUtilities`` and
now includes the TypeHandling Utility as well. That makes this package
fully independent from Flow and both Utilities make sense together.
