+++ Neos.ObjectUtilities/Classes/TYPO3/Flow/Reflection/Exception/PropertyNotAccessibleException.php
@@ -2,7 +2,7 @@
 namespace TYPO3\Flow\Reflection\Exception;
 
 /*
- * This file is part of the Neos.PropertyAccess package.
+ * This file is part of the Neos.ObjectUtilities package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
  *
