+++ Neos.ObjectUtilities/Classes/TYPO3/Flow/Reflection/ObjectAccess.php
@@ -2,7 +2,7 @@
 namespace TYPO3\Flow\Reflection;
 
 /*
- * This file is part of the Neos.PropertyAccess package.
+ * This file is part of the Neos.ObjectUtilities package.
  *
  * (c) Contributors of the Neos Project - www.neos.io
  *
