+++ Neos.ObjectUtilities/Readme.rst
@@ -1,5 +1,5 @@
 ----------------------------
-Flow Property Access Utility
+Neos Object Utilities
 ----------------------------
 
 .. note:: This repository is a **read-only subsplit** of a package that is part of the
@@ -15,10 +15,8 @@ If you want to contribute to the Flow framework, please have a look at
 https://github.com/neos/flow-development-collection - it is the repository
 used for development and all pull requests should go into it.
 
-
 Dependencies
 ------------
 
-This package depends on the TypeHandling utility of Flow, after that is split we can
-depend on that. Additionally the tests depend on Doctrine/ORM which is therefore
+The tests depend on Doctrine/ORM which is therefore
 a dev requirement requirement.
