+++ Neos.ObjectUtilities/Tests/Unit/Fixture/Model/EntityWithDoctrineProxy.php
@@ -1,5 +1,15 @@
 <?php
-namespace Neos\PropertyAccess\Tests\Unit\Fixture\Model;
+namespace Neos\ObjectUtilities\Tests\Unit\Fixture\Model;
+
+/*
+ * This file is part of the Neos.ObjectUtilities package.
+ *
+ * (c) Contributors of the Neos Project - www.neos.io
+ *
+ * This package is Open Source Software. For the full copyright and license
+ * information, please view the LICENSE file which was distributed with this
+ * source code.
+ */
 
 use Doctrine\ORM\Proxy\Proxy;
