+++ composer.json
@@ -23,7 +23,7 @@
         "neos/composer-plugin": "^1.0.1"
     },
     "replace": {
-        "neos/property-access": "self.version",
+        "neos/object-utilities": "self.version",
         "typo3/eel": "self.version",
         "typo3/flow": "self.version",
         "typo3/fluid": "self.version",
@@ -37,7 +37,7 @@
     },
     "autoload": {
         "psr-0": {
-            "TYPO3\\Flow\\Reflection": "Neos.PropertyAccess/Classes",
+            "TYPO3\\Flow\\Reflection": "Neos.ObjectUtilities/Classes",
             "TYPO3\\Eel": "TYPO3.Eel/Classes",
             "TYPO3\\Flow": "TYPO3.Flow/Classes",
             "TYPO3\\Fluid": "TYPO3.Fluid/Classes",
@@ -48,6 +48,14 @@
         "installer-name": "Framework",
         "neos": {
             "warning": "AUTOGENERATED FILE, ONLY MODIFY THE .composer.json IN THIS DIRECTORY AND RUN THE MERGE SCRIPT."
+        }
+    },
+    "require-dev": {
+        "doctrine/orm": "~2.4.0"
+    },
+    "autoload-dev": {
+        "psr-4": {
+            "Neos\\ObjectUtilities\\Tests\\": "Neos.ObjectUtilities/Tests"
         }
     }
 }
